refactor(shared): file-organization sweep over app/shared/models/resource
The empty note document is a constant, so it moves to the resource services' constants file, and the resource list sort key schema takes its own model file with the type it validates rather than riding beside ResourceListItem.
