+++ apps/web/app/store/resource/note/index.ts
@@ -1,4 +1,4 @@
-import { EMPTY_NOTE_DOC } from "#shared/models/resource/note/NoteResource";
+import { EMPTY_NOTE_DOC } from "#shared/services/resource/constants";
 import { createContentData } from "@/services/resource/createContentData";
 import { ResourceType } from "@esposter/db-schema";
