+++ apps/web/shared/models/db/resource/ReadDeletedResourcesInput.ts
@@ -1,7 +1,7 @@
 import type { z } from "zod";
 
 import { createOffsetPaginationParamsSchema } from "#shared/models/pagination/offset/OffsetPaginationParams";
-import { resourceListSortKeySchema } from "#shared/models/resource/ResourceListItem";
+import { resourceListSortKeySchema } from "#shared/models/resource/ResourceListSortKey";
 
 export const readDeletedResourcesInputSchema = createOffsetPaginationParamsSchema(resourceListSortKeySchema).prefault(
   {},
